pad on left for xlnet

+++ examples/run_xlnet_classifier.py
@@ -198,14 +198,17 @@ def main():
             list(filter(None, args.xlnet_model.split('/'))).pop(),
                         str(args.max_seq_length),
                         str(task_name)))
-        try:
+        if os.path.exists(cached_train_features_file):
+            logger.info("Loading train features for cache file %s", cached_train_features_file)
             with open(cached_train_features_file, "rb") as reader:
                 train_features = pickle.load(reader)
-        except:
+        else:
+            logger.info("No cache file at %s, preparing train features", cached_train_features_file)
             train_features = convert_examples_to_features(
                 train_examples, label_list, args.max_seq_length, tokenizer, output_mode,
                 cls_token_at_end=True, cls_token=tokenizer.CLS_TOKEN,
-                sep_token=tokenizer.SEP_TOKEN, cls_token_segment_id=2)
+                sep_token=tokenizer.SEP_TOKEN, cls_token_segment_id=2,
+                pad_on_left=True, pad_token_segment_id=4)
             if args.local_rank == -1 or torch.distributed.get_rank() == 0:
                 logger.info("  Saving train features into cached file %s", cached_train_features_file)
                 with open(cached_train_features_file, "wb") as writer:
@@ -344,14 +347,17 @@ def main():
             list(filter(None, args.xlnet_model.split('/'))).pop(),
                         str(args.max_seq_length),
                         str(task_name)))
-        try:
+        if os.path.exists(cached_eval_features_file):
+            logger.info("Loading eval features for cache file %s", cached_eval_features_file)
             with open(cached_eval_features_file, "rb") as reader:
                 eval_features = pickle.load(reader)
-        except:
+        else:
+            logger.info("No cache file at %s, preparing eval features", cached_eval_features_file)
             eval_features = convert_examples_to_features(
                 eval_examples, label_list, args.max_seq_length, tokenizer, output_mode,
                 cls_token_at_end=True, cls_token=tokenizer.CLS_TOKEN,
-                sep_token=tokenizer.SEP_TOKEN, cls_token_segment_id=2)
+                sep_token=tokenizer.SEP_TOKEN, cls_token_segment_id=2,
+                pad_on_left=True, pad_token_segment_id=4)
             if args.local_rank == -1 or torch.distributed.get_rank() == 0:
                 logger.info("  Saving eval features into cached file %s", cached_eval_features_file)
                 with open(cached_eval_features_file, "wb") as writer:
